// ECOT-Food-Corner/src/app/model/user.ts
export class User {
        firstName: string = '';
        lastName: string = '';
        dateofBirth: string = '';
        roomNumber: string = '';
        phoneNumber: string = '';
        email: string = '';
        password: string = '';
        confirmPassword: string = '';
        role: string = '';
        token: string = '';
}